eventpolls/serializers.py
from rest_framework import serializers

from .models import Event


class EventSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Event
        fields = ("name", "description", "start_date", "end_date")
